fix(billing): Fix all compilation errors
- CafManager: Cast long to Integer in setFolioActual
- ChileBillingProvider: Remove incorrect @Override annotation
- DteService: Convert Integer cantidad to BigDecimal before multiply
- All 6 compilation errors resolved

+++ backend/billing-service/src/main/java/com/poscl/billing/infrastructure/providers/chile/caf/CafManager.java
@@ -42,7 +42,7 @@ public class CafManager {
 
         // Incrementar folio
         long siguienteFolio = caf.getFolioActual() + 1;
-        caf.setFolioActual(siguienteFolio);
+        caf.setFolioActual((int) siguienteFolio);
 
         // Calcular folios restantes
         long foliosRestantes = caf.getFolioHasta() - siguienteFolio;
@@ -85,6 +85,6 @@ public class CafManager {
     public long getFoliosRestantes(UUID tenantId, TipoDte tipoDte) {
         return cafRepository.findActiveCafByTenantAndTipo(tenantId, tipoDte)
                 .map(caf -> caf.getFolioHasta() - caf.getFolioActual())
-                .orElse(0L);
+                .orElse(0).longValue();
     }
 }
